Split PickDetailView.spec.tsx into PickDetailView-tests/
The spec reached 681 lines, over the 600-line test-file cap, failing the
file-length CI check. Split into a PickDetailView-tests/ directory:
shared fixtures in helpers.ts, rendering/metadata/closed-state tests in
rendering.spec.tsx, and ranking/options/suggest/interaction tests in
interactions.spec.tsx. Each mock harness captures only the callbacks its
half reads. No test behavior changed.
